// src/app/modules/Course/course.interface.ts
import { Types } from "mongoose";

export type TPreRequisiteCourse = {
  course: Types.ObjectId;
  isDeleted: boolean;
};
export type TCourse = {
  title: string;
  prefix: string;
  code: number;
  credits: number;
  preRequisiteCourse: TPreRequisiteCourse[];
  isDeleted: boolean;
};

export type TCourseFaculties = {
  course: Types.ObjectId;
  faculties: Types.ObjectId[];
};